+++ src/main/java/racingcar/validation/RoundNumberValidator.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.validation;
 
 public class RoundNumberValidator {
