+++ src/test/java/racingcar/model/CarMakerTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.Car;
+import racingcar.validation.CarNameValidator;
 
 class CarMakerTest {
