+++ src/test/java/racingcar/model/CarNameValidatorTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.validation.CarNameValidator;
 
 class CarNameValidatorTest {
