+++ src/test/java/racingcar/model/RoundNumberValidatorTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.validation.RoundNumberValidator;
 
 class RoundNumberValidatorTest {
